move reusable CAL_9 code into liscal/diagnostic_plots.py
